only ints

+++ src/main/java/sep/fimball/general/data/Config.java
@@ -52,7 +52,7 @@ public class Config
         return pathToElements() + "/" + elementTypeId + pathElementToDataJson;
     }
 
-    public static String pathToElementImage(String elementTypeId, ImageLayer imageLayer, boolean canRotate, double rotation, boolean animation, String animationName, int animationId)
+    public static String pathToElementImage(String elementTypeId, ImageLayer imageLayer, boolean canRotate, int rotation, boolean animation, String animationName, int animationId)
     {
         String path = pathToElements() + "/" + elementTypeId + "/";
